feat(bookmark-group): 增加分组功能(WIP),并重构命令注册部分
