+++ src/commands/bookmark/index.ts
@@ -1,0 +1,3 @@
+export * from './group';
+export * from './view';
+export * from './base';
