+++ src/constants/bookmark.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_BOOKMARK_GROUP_ID = '-999999';
