+++ src/constants/configuration.ts
@@ -1,1 +1,3 @@
 export const DEFAULT_BOOKMARK_COLOR = '#0e69d8';
+
+export const DEFAULT_BOOKMARK_GROUP_COLOR = DEFAULT_BOOKMARK_COLOR;
