+++ src/providers/BookmarksTreeItem.ts
@@ -8,7 +8,10 @@ import {
 } from 'vscode';
 
 import BaseTreeItem from './BaseTreeItem';
-import {BookmarksGroupedByFileType} from '../types';
+import {
+  BookmarksGroupedByFileType,
+  BookmarksGroupedByCustomType,
+} from '../types';
 import {getLineInfoStrFromBookmark} from '../utils';
 import {CMD_GO_TO_SOURCE_LOCATION} from '../constants';
 import {ServiceManager} from '../services/ServiceManager';
@@ -23,7 +26,8 @@ export default class BookmarkTreeItem extends BaseTreeItem {
     | IBookmark
     | BookmarksGroupedByFileType
     | BookmarksGroupedByColorType
-    | BookmarksGroupedByWorkspaceType;
+    | BookmarksGroupedByWorkspaceType
+    | BookmarksGroupedByCustomType;
 
   private _sm: ServiceManager;
 
@@ -35,7 +39,8 @@ export default class BookmarkTreeItem extends BaseTreeItem {
       | IBookmark
       | BookmarksGroupedByFileType
       | BookmarksGroupedByColorType
-      | BookmarksGroupedByWorkspaceType,
+      | BookmarksGroupedByWorkspaceType
+      | BookmarksGroupedByCustomType,
     sm: ServiceManager,
   ) {
     super(label, collapsibleState, contextValue);
@@ -56,6 +61,9 @@ export default class BookmarkTreeItem extends BaseTreeItem {
       this._resolveFileOverview();
       this.label = _meta.fileName;
       this.description = workspace.asRelativePath(_meta.fileId);
+    } else if (this.contextValue === 'custom') {
+      this.label = label;
+      this.iconPath = ThemeIcon.Folder;
     } else {
       this.command = {
         title: l10n.t('Jump to bookmark position'),
