+++ src/stores/index.ts
@@ -2,3 +2,6 @@ export * from './bookmark';
 export * from './configure';
 export * from './decoration';
 export * from './custom';
+export * from './bookmark';
+export * from './bookmark-group';
+export * from './bookmark-store';
