+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './common';
 export * from './code';
 export * from './universal';
+export * from './bookmarks';
