+++ tsconfig.json
@@ -8,6 +8,7 @@
     "strict": true,
     "noImplicitAny": true,
     "esModuleInterop": true,
+    "experimentalDecorators": true,
     /* enable all strict type-checking options */
     /* Additional Checks */
     // "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
